refactor(tour): modularize shepherd tour into components, utils and styles
