+++ src/styles/components/Tour.css
@@ -1,0 +1,29 @@
+.shepherd-button-secondary {
+ background-color:#6c757d !important;
+}
+
+.shepherd-button {
+ background-color:#050c40 !important;
+ color:white !important;
+}
+
+.tour-header{
+ display:flex;
+ justify-content:space-between;
+ border-bottom:1px solid #eee;
+ padding-bottom:4px;
+}
+
+.tour-title{
+ font-weight:600;
+ font-size:13px;
+}
+
+.tour-progress{
+ font-size:12px;
+ color:#777;
+}
+
+.tour-body{
+ font-size:14px;
+}
