+++ src/types/testing-library-react.d.ts
@@ -1,0 +1,6 @@
+// Provide a bare-bones module declaration in case TypeScript cannot locate the
+// built-in types from the package. The proper types are shipped with
+// `@testing-library/react` so this file is just a fallback to silence editor
+// diagnostics.
+
+declare module "@testing-library/react";
